middleware with sweden connect pkcs#11 hsm
This project is the deployment version of the German eIDAS middleware project version 1.1.0 which adds our SAML fixes and adds Sweden Connect PKCS#11 implementation as means to provide the SAML signing and encryption key in PKCS#11 HSM.

This implementation is keeping the CVC keys in software. The CVC keys can't be used to impersonate the eIDAS node, only to request authentication. For this reason the CVC keys is not as important to protect as the SAML keys.